Filter: Refactor based on reviews
