+++ app/src/main/java/com/ayhancavdar/nexusbanking/core/navigation/AppNavHost.kt
@@ -15,7 +15,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import com.ayhancavdar.nexusbanking.features.accountDetails.accountDetails
 import com.ayhancavdar.nexusbanking.features.accounts.presentation.navigation.accounts
-import com.ayhancavdar.nexusbanking.features.filter.filter
+import com.ayhancavdar.nexusbanking.features.filter.navigation.filter
 import com.ayhancavdar.nexusbanking.features.login.presentation.navigation.login
 import com.ayhancavdar.nexusbanking.features.otp.presentation.navigation.otp
 import com.ayhancavdar.nexusbanking.features.splash.splash
