+++ app/src/main/java/com/ayhancavdar/nexusbanking/core/navigation/NexusBankingRoute.kt
@@ -9,6 +9,7 @@
 
 package com.ayhancavdar.nexusbanking.core.navigation
 
+import com.ayhancavdar.nexusbanking.features.filter.navigation.FilterParametersNavType
 import com.ayhancavdar.nexusbanking.features.filter.state.FilterParameters
 import kotlin.reflect.typeOf
 import kotlinx.serialization.Serializable
